Больше, больше, больше numpy :D

+++ work_with_file.py
@@ -4,7 +4,7 @@ import h5py
 def values_from_file(path: str = None) -> np.ndarray:
     try:
         with open(path) as f:
-            return np.loadtxt(f, dtype=np.int8)
+            return np.loadtxt(f)
             # return values if len(values) != 0 else None
     except FileNotFoundError:
         return None
